add ~features_from_matches parameter to control how features a created
`~features_from_matches` (`bool`, default: `true`): if set to true, features
appearing in a `candidate_matches` message are assumed to be currently
tracked, and are added to the features-person graph. If set to false, only
features published on `/humans/*/tracked` are considered tracked, and
`candidate_matches` for untracked features will be ignored.